Fixed adding grade bug, added ability to add comments to grades
Comments currently do not have a max buffer size, which will need to be
fixed later. The "Add Grade" bug would occur after the user added the
grade information and clicked "OK". It would occur because, although a
new course view was being added in GBTreeListener class, the previous
view and graphical data was not removed from the mainpanel. So even
though the mainpanel was revalidated and appeared to have another grade
in under the user selected category, there was only the initialized
amount of text boxes resulting in overlap when the mainpanel was
painted.